pyproject.toml
[tool.poetry]
name = "smsgateway"
version = "0.1.0"
description = "SMS Gateway Backend Service"
authors = ["Alexandr Chikur <alexandr.chikur@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
django-rest-framework = "^0.1.0"
django-debug-toolbar = "^3.8.1"
dj-database-url = "^1.2.0"
psycopg2 = "^2.9.5"
djoser = "^2.1.0"
phonenumbers = "^8.13.3"
drf-yasg = "^1.21.4"
djangochannelsrestframework = "^1.1.0"
uvicorn = "^0.20.0"
websockets = "^10.4"
channels-redis = "^4.0.0"
whitenoise = "^6.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
